Se eliminaron archivos innecesarios, se hacen pruebas para asegurar las rutas.
