+++ src/App.js
@@ -6,9 +6,20 @@ import GlobalStyles from 'src/components/GlobalStyles';
 import 'src/mixins/chartjs';
 import theme from 'src/theme';
 import routes from 'src/routes';
+import { useAuth0, isAuthenticated } from "@auth0/auth0-react";
 
 const App = () => {
   const routing = useRoutes(routes);
+
+  const { isLoading, error } = useAuth0();
+
+  if (error) {
+    return <div>Oops... {error.message}</div>;
+  }
+
+  if (isLoading) {
+    return <h3>Cargando</h3> // <Loading />;
+  }
 
   return (
     <ThemeProvider theme={theme}>
