+++ src/index.js
@@ -4,6 +4,15 @@ import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import { Auth0Provider } from '@auth0/auth0-react';
 import config from './auth_config.json';
+import history from "./utils/history";
+
+const onRedirectCallback = (appState) => {
+  history.push(
+    appState && appState.returnTo
+      ? appState.returnTo
+      : window.location.pathname
+  );
+};
 
 ReactDOM.render(
   <Auth0Provider
@@ -12,8 +21,9 @@ ReactDOM.render(
     audience={config.audience}
     scope="read:current_user update:current_user_metadata"
     redirectUri={window.location.origin}
+    onRedirectCallback={onRedirectCallback}
   >
-    <BrowserRouter>
+    <BrowserRouter history={history}>
       <App />
     </BrowserRouter>
   </Auth0Provider>,
