+++ src/views/auth/LoginView.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import * as Yup from 'yup';
-import { Formik } from 'formik';
 import {
   Box,
   Button,
@@ -11,9 +9,7 @@ import {
   makeStyles
 } from '@material-ui/core';
 import Page from 'src/components/Page';
-import LoginButton from '../../components/LoginButton';
-import LogoutButton from '../../components/LogoutButton';
-import Profile from '../../components/Profile';
+import { useAuth0 } from '@auth0/auth0-react';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -25,6 +21,7 @@ const useStyles = makeStyles(theme => ({
 }));
 
 const LoginView = () => {
+  const { loginWithRedirect } = useAuth0();
   const classes = useStyles();
   const navigate = useNavigate();
 
@@ -36,93 +33,28 @@ const LoginView = () => {
         height="100%"
         justifyContent="center"
       >
-        <div>
-          <Profile/>
-        </div>
         <Container maxWidth="sm">
-        <LoginButton/>
-        <LogoutButton/>
-          <Formik
-            initialValues={{
-              email: 'demo@devias.io',
-              password: 'Password123'
-            }}
-            validationSchema={Yup.object().shape({
-              email: Yup.string()
-                .email('Must be a valid email')
-                .max(255)
-                .required('Email is required'),
-              password: Yup.string()
-                .max(255)
-                .required('Password is required')
-            })}
-            onSubmit={() => {
-              navigate('/app/upload', { replace: true });
-            }}
-          >
-            {({
-              errors,
-              handleBlur,
-              handleChange,
-              handleSubmit,
-              isSubmitting,
-              touched,
-              values
-            }) => (
-              <form onSubmit={handleSubmit}>
-                <Box mb={3}>
-                  <Typography color="textPrimary" variant="h2">
-                    Inicio de sesión
-                  </Typography>
-                  <Typography
-                    color="textSecondary"
-                    gutterBottom
-                    variant="body2"
-                  >
-                    Inicia sesión en la plataforma de people analytics
-                  </Typography>
-                </Box>
-                <TextField
-                  error={Boolean(touched.email && errors.email)}
-                  fullWidth
-                  helperText={touched.email && errors.email}
-                  label="Dirección de correo electrónico"
-                  margin="normal"
-                  name="email"
-                  onBlur={handleBlur}
-                  onChange={handleChange}
-                  type="email"
-                  value={values.email}
-                  variant="outlined"
-                />
-                <TextField
-                  error={Boolean(touched.password && errors.password)}
-                  fullWidth
-                  helperText={touched.password && errors.password}
-                  label="Contraseña"
-                  margin="normal"
-                  name="password"
-                  onBlur={handleBlur}
-                  onChange={handleChange}
-                  type="password"
-                  value={values.password}
-                  variant="outlined"
-                />
-                <Box my={2}>
-                  <Button
-                    color="primary"
-                    disabled={isSubmitting}
-                    fullWidth
-                    size="large"
-                    type="submit"
-                    variant="contained"
-                  >
-                    Ingresar
-                  </Button>
-                </Box>
-              </form>
-            )}
-          </Formik>
+          <Box mb={3}>
+            <Typography color="textPrimary" variant="h2">
+              Bienvenido
+            </Typography>
+            <Typography color="textSecondary" gutterBottom variant="body2">
+              Inicia sesión en la plataforma de people analytics
+            </Typography>
+          </Box>
+
+          <Box my={2}>
+            <Button
+              color="primary"
+              fullWidth
+              size="large"
+              type="submit"
+              variant="contained"
+              onClick={() => loginWithRedirect()}
+            >
+              Ingresar
+            </Button>
+          </Box>
         </Container>
       </Box>
     </Page>
