--- src/views/reports/DashboardView/LatestOrders.js
@@ -1,185 +1,0 @@
-import React, { useState } from 'react';
-import clsx from 'clsx';
-import moment from 'moment';
-import { v4 as uuid } from 'uuid';
-import PerfectScrollbar from 'react-perfect-scrollbar';
-import PropTypes from 'prop-types';
-import {
-  Box,
-  Button,
-  Card,
-  CardHeader,
-  Chip,
-  Divider,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableRow,
-  TableSortLabel,
-  Tooltip,
-  makeStyles
-} from '@material-ui/core';
-import ArrowRightIcon from '@material-ui/icons/ArrowRight';
-
-const data = [
-  {
-    id: uuid(),
-    ref: 'CDD1049',
-    amount: 30.5,
-    customer: {
-      name: 'Juana Martinez'
-    },
-    createdAt: 1602038931000 ,
-    status: 'riesgo académico'
-  },
-  {
-    id: uuid(),
-    ref: 'CDD1048',
-    amount: 25.1,
-    customer: {
-      name: 'Gabriel Ramirez'
-    },
-    createdAt: 1602038931000,
-    status: 'psicoemocional'
-  },
-  {
-    id: uuid(),
-    ref: 'CDD1047',
-    amount: 10.99,
-    customer: {
-      name: 'Alexandra Lopez'
-    },
-    createdAt: 1601025931000,
-    status: 'riesgo economico'
-  },
-  {
-    id: uuid(),
-    ref: 'CDD1046',
-    amount: 96.43,
-    customer: {
-      name: 'Daniel Perez'
-    },
-    createdAt: 1600925931000,
-    status: 'riesgo economico'
-  },
-  {
-    id: uuid(),
-    ref: 'CDD1045',
-    amount: 32.54,
-    customer: {
-      name: 'Gilberto Tobon'
-    },
-    createdAt: 1600925931000,
-    status: 'seguimiento'
-  },
-  {
-    id: uuid(),
-    ref: 'CDD1044',
-    amount: 16.76,
-    customer: {
-      name: 'Adrian Cano'
-    },
-    createdAt: 1600825931000,
-    status: 'seguimiento'
-  }
-];
-
-const useStyles = makeStyles(() => ({
-  root: {},
-  actions: {
-    justifyContent: 'flex-end'
-  }
-}));
-
-const LatestOrders = ({ className, ...rest }) => {
-  const classes = useStyles();
-  const [orders] = useState(data);
-
-  return (
-    <Card
-      className={clsx(classes.root, className)}
-      {...rest}
-    >
-      <CardHeader title="Mis Ultimos Casos" />
-      <Divider />
-      <PerfectScrollbar>
-        <Box minWidth={800}>
-          <Table>
-            <TableHead>
-              <TableRow>
-                <TableCell>
-                  ID Estudiante
-                </TableCell>
-                <TableCell>
-                  Nombre
-                </TableCell>
-                <TableCell sortDirection="desc">
-                  <Tooltip
-                    enterDelay={300}
-                    title="Sort"
-                  >
-                    <TableSortLabel
-                      active
-                      direction="desc"
-                    >
-                      Ultimo contacto
-                    </TableSortLabel>
-                  </Tooltip>
-                </TableCell>
-                <TableCell>
-                  Estado
-                </TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {orders.map((order) => (
-                <TableRow
-                  hover
-                  key={order.id}
-                >
-                  <TableCell>
-                    {order.ref}
-                  </TableCell>
-                  <TableCell>
-                    {order.customer.name}
-                  </TableCell>
-                  <TableCell>
-                    {moment(order.createdAt).format('DD/MM/YYYY')}
-                  </TableCell>
-                  <TableCell>
-                    <Chip
-                      color="primary"
-                      label={order.status}
-                      size="small"
-                    />
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </Box>
-      </PerfectScrollbar>
-      <Box
-        display="flex"
-        justifyContent="flex-end"
-        p={2}
-      >
-        <Button
-          color="primary"
-          endIcon={<ArrowRightIcon />}
-          size="small"
-          variant="text"
-        >
-          Ver Todos
-        </Button>
-      </Box>
-    </Card>
-  );
-};
-
-LatestOrders.propTypes = {
-  className: PropTypes.string
-};
-
-export default LatestOrders;
