--- src/views/reports/DashboardView/TotalProfit.js
@@ -1,70 +1,0 @@
-import React from 'react';
-import clsx from 'clsx';
-import PropTypes from 'prop-types';
-import {
-  Avatar,
-  Card,
-  CardContent,
-  Grid,
-  Typography,
-  makeStyles,
-  colors
-} from '@material-ui/core';
-import AssistantPhoto from '@material-ui/icons/AssistantPhoto';
-
-const useStyles = makeStyles(() => ({
-  root: {
-    height: '100%'
-  },
-  avatar: {
-    backgroundColor: colors.indigo[600],
-    height: 56,
-    width: 56
-  }
-}));
-
-const TotalProfit = ({ className, ...rest }) => {
-  const classes = useStyles();
-
-  return (
-    <Card
-      className={clsx(classes.root, className)}
-      {...rest}
-    >
-      <CardContent>
-        <Grid
-          container
-          justify="space-between"
-          spacing={3}
-        >
-          <Grid item>
-            <Typography
-              color="textSecondary"
-              gutterBottom
-              variant="h6"
-            >
-              ESTUDIANTES INTERVENIDOS
-            </Typography>
-            <Typography
-              color="textPrimary"
-              variant="h3"
-            >
-              1,523
-            </Typography>
-          </Grid>
-          <Grid item>
-            <Avatar className={classes.avatar}>
-              <AssistantPhoto/>
-            </Avatar>
-          </Grid>
-        </Grid>
-      </CardContent>
-    </Card>
-  );
-};
-
-TotalProfit.propTypes = {
-  className: PropTypes.string
-};
-
-export default TotalProfit;
