--- src/views/reports/DashboardView/TrafficByDevice.js
@@ -1,150 +1,0 @@
-import React from 'react';
-import clsx from 'clsx';
-import PropTypes from 'prop-types';
-import { Doughnut } from 'react-chartjs-2';
-import {
-  Box,
-  Card,
-  CardContent,
-  CardHeader,
-  Divider,
-  Typography,
-  colors,
-  makeStyles,
-  useTheme
-} from '@material-ui/core';
-import LaptopMacIcon from '@material-ui/icons/CheckCircleOutline';
-import PhoneIcon from '@material-ui/icons/NotificationsNone';
-import TabletIcon from '@material-ui/icons/Warning';
-
-const useStyles = makeStyles(() => ({
-  root: {
-    height: '100%'
-  }
-}));
-
-const TrafficByDevice = ({ className, ...rest }) => {
-  const classes = useStyles();
-  const theme = useTheme();
-
-  const data = {
-    datasets: [
-      {
-        data: [30, 60, 10],
-        backgroundColor: [
-          colors.indigo[500],
-          colors.red[600],
-          colors.orange[600]
-        ],
-        borderWidth: 8,
-        borderColor: colors.common.white,
-        hoverBorderColor: colors.common.white
-      }
-    ],
-    labels: ['Ok', 'Atencion', 'Riesgo']
-  };
-
-  const options = {
-    animation: false,
-    cutoutPercentage: 80,
-    layout: { padding: 0 },
-    legend: {
-      display: false
-    },
-    maintainAspectRatio: false,
-    responsive: true,
-    tooltips: {
-      backgroundColor: theme.palette.background.default,
-      bodyFontColor: theme.palette.text.secondary,
-      borderColor: theme.palette.divider,
-      borderWidth: 1,
-      enabled: true,
-      footerFontColor: theme.palette.text.secondary,
-      intersect: false,
-      mode: 'index',
-      titleFontColor: theme.palette.text.primary
-    }
-  };
-
-  const devices = [
-    {
-      title: 'Buen Desempeño',
-      value: 20,
-      icon: LaptopMacIcon,
-      color: colors.indigo[500]
-    },
-    {
-      title: 'En Alerta Predictiva',
-      value: 10,
-      icon: PhoneIcon,
-      color: colors.orange[600]
-    },
-    {
-      title: 'En Alerta Latente',
-      value: 60,
-      icon: TabletIcon,
-      color: colors.red[600]
-    }
-
-  ];
-
-  return (
-    <Card
-      className={clsx(classes.root, className)}
-      {...rest}
-    >
-      <CardHeader title="Riesgo de Deserción" />
-      <Divider />
-      <CardContent>
-        <Box
-          height={300}
-          position="relative"
-        >
-          <Doughnut
-            data={data}
-            options={options}
-          />
-        </Box>
-        <Box
-          display="flex"
-          justifyContent="center"
-          mt={2}
-        >
-          {devices.map(({
-            color,
-            icon: Icon,
-            title,
-            value
-          }) => (
-            <Box
-              key={title}
-              p={1}
-              textAlign="center"
-            >
-              <Icon color="action" />
-              <Typography
-                color="textPrimary"
-                variant="body1"
-              >
-                {title}
-              </Typography>
-              <Typography
-                style={{ color }}
-                variant="h2"
-              >
-                {value}
-                %
-              </Typography>
-            </Box>
-          ))}
-        </Box>
-      </CardContent>
-    </Card>
-  );
-};
-
-TrafficByDevice.propTypes = {
-  className: PropTypes.string
-};
-
-export default TrafficByDevice;
