+++ src/views/upload/UploadView/index.js
@@ -3,7 +3,6 @@ import { Container, makeStyles } from '@material-ui/core';
 import Page from 'src/components/Page';
 import Upload from './Upload';
 
-
 const useStyles = makeStyles(theme => ({
   root: {
     backgroundColor: theme.palette.background.dark,
